re-organize the repo, move classes to individual folders. Create class for packaged DX equipment.
