+++ .gitignore
@@ -1,2 +1,3 @@
 *.cpython-37.pyc
 /copper/__pycache__/
+dev
