+++ copper/__init__.py
@@ -1,6 +1,5 @@
 __author__ = """Jeremy Lerond"""
 __email__ = "jeremy.lerond@pnnl.gov"
 
-from copper.copper import *
-
-__all__ = ["Copper"]
+from copper.chiller import *
+from copper.pkg_dx import *
